# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Estados",
        "description": "API para gestión de estados",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://uptyab.test/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/ubicaciones": {
            "get": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Obtener lista de ubicaciones",
                "operationId": "f02b0c7bb15c07620cb90f5d02e85c04",
                "responses": {
                    "200": {
                        "description": "Lista de ubicaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ubicacion"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Crear una nueva ubicación",
                "operationId": "6de080f6bce2ff5ee12f66b8582803fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comunidad_id",
                                    "nombre"
                                ],
                                "properties": {
                                    "comunidad_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Ubicación Principal"
                                    },
                                    "calle": {
                                        "type": "string",
                                        "example": "Calle Principal"
                                    },
                                    "avenida": {
                                        "type": "string",
                                        "example": "Avenida Bolívar"
                                    },
                                    "referencia": {
                                        "type": "string",
                                        "example": "Frente a la plaza"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ubicación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ubicacion"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/ubicaciones/{id}": {
            "get": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Obtener una ubicación específica",
                "operationId": "5fcf8be8fa7dd2ffd3b40f322ff96fb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ubicación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ubicación encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ubicacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ubicación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Actualizar una ubicación existente",
                "operationId": "e4538fce84aecad890f922096eb62172",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ubicación a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "comunidad_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Ubicación Actualizada"
                                    },
                                    "calle": {
                                        "type": "string",
                                        "example": "Calle Actualizada"
                                    },
                                    "avenida": {
                                        "type": "string",
                                        "example": "Avenida Actualizada"
                                    },
                                    "referencia": {
                                        "type": "string",
                                        "example": "Referencia actualizada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ubicación actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ubicacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ubicación no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Eliminar una ubicación",
                "operationId": "b247926f51daa193f72b8fa31c88174a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ubicación a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ubicación eliminada exitosamente"
                    },
                    "404": {
                        "description": "Ubicación no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ubicacion": {
                "title": "Ubicacion",
                "description": "Modelo que representa una ubicación",
                "properties": {
                    "id": {
                        "description": "ID de la ubicación",
                        "type": "integer",
                        "format": "int64"
                    },
                    "comunidad_id": {
                        "description": "ID de la comunidad a la que pertenece",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "description": "Nombre de la ubicación",
                        "type": "string"
                    },
                    "calle": {
                        "description": "Calle de la ubicación",
                        "type": "string",
                        "nullable": true
                    },
                    "avenida": {
                        "description": "Avenida de la ubicación",
                        "type": "string",
                        "nullable": true
                    },
                    "referencia": {
                        "description": "Referencia de la ubicación",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}